// src/api/generated/model/internalHandlerAuditListItem.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalHandlerAuditListItemDetail } from './internalHandlerAuditListItemDetail';

export interface InternalHandlerAuditListItem {
  action?: string;
  actor_role?: string;
  actor_user_id?: string;
  actor_username?: string;
  created_at?: string;
  detail?: InternalHandlerAuditListItemDetail;
  error_message?: string;
  id?: string;
  method?: string;
  path?: string;
  remote_ip?: string;
  resource_id?: string;
  resource_type?: string;
  result?: string;
  risk_level?: string;
  trace_id?: string;
}