// src/api/generated/model/internalHandlerAuditListResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalHandlerAuditListItem } from './internalHandlerAuditListItem';
import type { ZvasPkgHttpxPagination } from './zvasPkgHttpxPagination';

export interface InternalHandlerAuditListResponse {
  code?: number;
  data?: InternalHandlerAuditListItem[];
  message?: string;
  pagination?: ZvasPkgHttpxPagination;
  trace_id?: string;
}